+++ assets/javascripts/kitten/components/carousel/carousel/carousel.js
@@ -65,6 +65,7 @@ class CarouselBase extends Component {
   static defaultProps = {
     hidePaginationOnMobile: false,
     showOtherPages: false,
+    pagesClassName: null,
     paginationPosition: {
       default: 'right',
       fromM: 'bottom',
@@ -79,6 +80,7 @@ class CarouselBase extends Component {
     viewportIsXSOrLess: PropTypes.bool.isRequired,
     hidePaginationOnMobile: PropTypes.bool,
     showOtherPages: PropTypes.bool,
+    pagesClassName: PropTypes.string,
     paginationPosition: PropTypes.shape({
       default: propTypesPositions,
       fromXxs: propTypesPositions,
@@ -164,6 +166,7 @@ class CarouselBase extends Component {
       viewportIsXSOrLess,
       viewportIsMOrLess,
       showOtherPages,
+      pagesClassName,
     } = this.props
     const { indexPageVisible, numColumns, numPages } = this.state
     const itemMarginBetween = getMarginBetweenAccordingToViewport(
@@ -183,6 +186,7 @@ class CarouselBase extends Component {
         onResizeInner={this.onResizeInner}
         goToPage={this.goToPage}
         showOtherPages={showOtherPages}
+        pagesClassName={pagesClassName}
       />
     )
   }